rendering problem

+++ app/Http/Controllers/Api/Artists/DeleteController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api\Artists;
 use App\Http\Controllers\Controller;
 use App\Models\Artist;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 
 class DeleteController extends Controller
 {
@@ -12,8 +13,10 @@ class DeleteController extends Controller
     {
 
         if ($artist = Artist::fetchBySlug($slug)) {
-            $oldAvatar = public_path('images/artists/') . $artist->avatar;
-            File::delete($oldAvatar);
+            if ($artist->avatar != 'default.jpg') {
+                $oldAvatar = public_path('images/artists/') . $artist->avatar;
+                File::delete($oldAvatar);
+            }
             $artist->delete();
             return response()->json('artist deleted');
         }
